+++ source/interface/Font.cpp
@@ -23,13 +23,13 @@ Font::Font() : Sprite("graphics/interface/font.png", 8, 16) {
 Font::~Font() {
 }
 
-void Font::drawChar(s16 x, s16 y, char c) {
+void Font::drawChar(float x, float y, char c) {
 	if(c >= 32) {
 		drawFrame(x, y, c - 29);
 	}
 }
 
-void Font::drawString(s16 x, s16 y, std::string str) {
+void Font::drawString(float x, float y, std::string str) {
 	for(u16 i = 0 ; i < str.length() ; i++) {
 		drawChar(x + (i * m_frameWidth), y, str[i]);
 	}
